added one thing that should increase speed by about 3 times

+++ semtag-core/src/main/java/org/semtag/model/Item.java
@@ -10,7 +10,7 @@ public class Item {
      * Constructs an item from a given ID.
      * @param itemId
      */
-    public Item(int itemId) {
+    public Item(long itemId) {
         this.itemId = String.valueOf(itemId);
     }
